Bill S2371 version original - Jamaal Bailey
BILL NUMBER: S2371 SPONSOR: BAILEY TITLE OF BILL: An act to amend the penal law, in relation to including solicitation of certain crimes as a specified offense for the purposes of a hate crime PURPOSE: The purpose of the bill is to add criminal solicitation and facilitation to the hate crimes statute. SUMMARY OF SPECIFIC PROVISIONS: Section one amends the definition of "specified offense" to include solicitation and facilitation of the crimes contained therein.