Bill S2468 version original - Kevin S. Parker
BILL NUMBER: S2468 SPONSOR: PARKER TITLE OF BILL: An act in relation to maintaining the continued viability of the state's existing large-scale, renewable energy resources PURPOSE: To amend the competitive tier 2 program within the Clean Energy Standard and ensure that the state's existing renewable resources are maintained and continue to contribute to the state's renewable energy goals. SUMMARY OF PROVISIONS: Section one of the bill provides the Legislative findings and intent of the bill.