Bill S2390 version original - John Liu
BILL NUMBER: S2390 SPONSOR: LIU TITLE OF BILL: An act to amend the general business law, in relation to the use of biometric data for marketing purposes PURPOSE OR GENERAL IDEA OF BILL: Relates to the use of biometric data for marketing purposes SUMMARY OF SPECIFIC PROVISIONS: Adds Section 399-dddd of the general business law. Section 1 subdivision (a) defines' private entity* to mean any individual.