Bill S2415 version original - Jamaal Bailey
BILL NUMBER: S2415 SPONSOR: BAILEY TITLE OF BILL: An act to amend the agriculture and markets law, in relation to estab- lishing the office of urban agriculture PURPOSE OR GENERAL IDEA: The purpose of this legislation is to establish an Office of Urban Agri- culture to assist with and promote various forms of urban farming throughout the state. SUMMARY OF SPECIFIC PROVISIONS: Section 1 Amends the heading of article 2-c of the agriculture and markets law to read, "Office of Urban Agriculture.