Bill S2375 version original - Roxanne J. Persaud
BILL NUMBER: S2375 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the tax law and the election law, in relation to requir- ing tax preparation software to provide a link to voter registration websites PURPOSE OR GENERAL IDEA OF BILL: This bill would be known as the "Representation through Taxation" bill. It would allow individuals the opportunity to register to vote on tax return software when doing their taxes. New York consistently ranks in the low 30 to 40 percent in voter turnout.