Bill S2374 version original - Roxanne J. Persaud
BILL NUMBER: S2374 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the public health law, in relation to out-of-network treatment disclosure for certain surgical procedures PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to amend the public health law, in relation to out-of-network treatment disclosure for certain surgical procedures. SUMMARY OF SPECIFIC PROVISIONS: Section 1.