Bill S2402 version original - Zellnor Myrie
BILL NUMBER: S2402 SPONSOR: MYRIE TITLE OF BILL: An act to amend the tax law, in relation to rebates on stock transfer tax paid; and to amend the state finance law, in relation to the funds of the stock transfer tax fund and the dedicated infrastructure invest- ment fund PURPOSE: Since at least 1915, New York has imposed a tax on the sale of securi- ties. Nonetheless, the securities industry flourished in the State.