Bill S2471 version original - John Liu
BILL NUMBER: S2471 SPONSOR: LIU TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring photographs of all parking violations PURPOSE OR GENERAL IDEA OF BILL: To require that purchases of equipment and technology be able to support that photographic evidence of parking violations. Also, to require that photographic evidence of parking violations be available and kept on record for all parking violation summons issued.