Bill S2355A version original - Simcha Felder
BILL NUMBER: S2355 SPONSOR: FELDER TITLE OF BILL: An act to amend the family court act, in relation to the hours of opera- tion for family courts; and providing for the repeal of such provisions upon expiration thereof PURPOSE OR GENERAL IDEA OF BILL: This bill would create a pilot program for family court in New York City to remain open until midnight one day a week for 3 boroughs. SUMMARY OF PROVISIONS: 1.