Bill S2413C version original - Jamaal Bailey
BILL NUMBER: S2413 SPONSOR: BAILEY TITLE OF BILL: An act to amend the general municipal law, in relation to providing for an annual adjustment for reimbursements for certain veterans funerals PURPOSE OR GENERAL IDEA OF BILL: The New York State funeral director association will have the capacity to adjust the funds that help families of deceased veterans pay for funerals. SUMMARY OF SPECIFIC PROVISIONS: Section 1.