Bill S2383 version original - Zellnor Myrie
BILL NUMBER: S2383 SPONSOR: MYRIE TITLE OF BILL: An act to amend to the election law, in relation to the administration of challenge oaths to voter applicants PURPOSE: The purpose of the bill is to ensure that voters are challenged in good faith. SUMMARY OF PROVISIONS: Section 1 amends section 8-502 of the election law which requires that permits the inspector or the clerk to challenge an applicant if they know or suspect the person is not entitled to vote in the district.