Bill S2411 version original - Jamaal Bailey
BILL NUMBER: S2411 SPONSOR: BAILEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to preserving the ability to appeal a violation of a defendant's right to make a statement personally in his or her own behalf at sentencing PURPOSE: The purpose of this bill is to ensure that a defendant retains his or her right to speak to the court at sentencing. SUMMARY OF PROVISIONS: Section one of this bill amends subdivision 1 of section 380.