Bill S2501 version original - Anthony H. Palumbo
BILL NUMBER: S2501 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the penal law, in relation to consecutive terms of imprisonment PURPOSE: When an offender commits two or more felonies and one involves possession or use of illegal firearm, this bill would require that the sentences run consecutively. SUMMARY OF PROVISIONS: Section 1. Section 70.25 of the penal law is amended by adding a new subdivision 2-h.