Bill S2421 version original - Brad Hoylman-Sigal
BILL NUMBER: S2421 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the penal law, in relation to electronic dart guns and electronic stun guns PURPOSE: The purpose of this bill is to regulate the sale, possession, and use of electronic dart guns and electronic stun guns in New York State. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill amends subdivision a of Section 265.20 of the Penal Law to add two new paragraphs 17 and 18.