Bill S2414 version original - Jamaal Bailey
BILL NUMBER: S2414 SPONSOR: BAILEY TITLE OF BILL: An act to authorize and direct the commissioner of agriculture and markets to conduct a study on vertical farming PURPOSE: To examine the effects vertical farming would have on access to locally grown food, job creation and educational opportunities, and the impacts on the environment. SUMMARY OF PROVISIONS: Section 1, subsection 1: Directs the commissioner of agriculture and markets to conduct a study on vertical farming. Subsection 2.