Bill S2401 version original - Zellnor Myrie
BILL NUMBER: S2401 SPONSOR: MYRIE TITLE OF BILL: An act to repeal chapter 834 of the laws of 1940, relating to the removal of certain policemen PURPOSE: To repeal chapter 834 of the laws of 1940, which limits the ability of localities to provide for independent oversight of police. SUMMARY OF PROVISIONS: Section 1 repeals chapter 834 of the laws of 1940 relating to certain policemen. Section 2 provides for the effective date.