Bill S2418 version original - Leroy Comrie
BILL NUMBER: S2418 SPONSOR: COMRIE TITLE OF BILL: An act to amend the executive law, in relation to expanding remedies for violations of New York state uniform fire prevention and building code PURPOSE: This bill increases fines for building code violations which creates an immediate threat to the safety and welfare of the building occupants. SUMMARY OF PROVISIONS: Section 1.