Bill S2423 version original - Brad Hoylman-Sigal
BILL NUMBER: S2423 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the executive law, in relation to parole eligibility for certain incarcerated persons aged fifty-five or older SUMMARY OF PROVISIONS: Section 1 of the bill adds new subdivisions 18 and 19 to section 259-c of the Executive Law ("State board of parole; functions, powers and duties"). -Sub.