Bill S2372 version original - Jamaal Bailey
BILL NUMBER: S2372 SPONSOR: BAILEY TITLE OF BILL: An act to amend the private housing finance law, in relation to requir- ing the publication of the number of veterans on certain housing waiting lists PURPOSE OR GENERAL IDEA OF BILL: The bill makes the number of veterans, or a surviving spouse of a veter- an, available on the waiting list for the Mitchell7Lama housing applica- tions. SUMMARY OF SPECIFIC PROVISIONS: Section 1.