Bill S2412 version original - Jamaal Bailey
BILL NUMBER: S2412 SPONSOR: BAILEY TITLE OF BILL: An act to amend the social services law, in relation to transportation management brokers PURPOSE: To eliminate from law the broker involvement in ambulette transportation under Medicaid. SUMMARY OF PROVISIONS: Section 1. Subdivision 4 of section 365-h of the social services law, as amended by section 2 of part LL of chapter 56 of the laws of 2020. Section 2.