Bill S2408 version original - Jamaal Bailey
BILL NUMBER: S2408 SPONSOR: BAILEY TITLE OF BILL: An act to amend the correction law, in relation to establishing a pilot program for the use of body-worn cameras on certain correction officers SUMMARY OF PROVISIONS: Section 1 of the bill creates a new section 150 of the correction law providing for a body camera pilot program in a state correctional facil- ity. Section 2 of the bill is the effective date.