Bill S2524 version original - Pamela Helming
BILL NUMBER: S2524 SPONSOR: HELMING TITLE OF BILL: An act to amend the penal law, in relation to the promotion of prison contraband PURPOSE: To enhance the penalties for introducing or possessing dangerous contra- band in New York State's prisons. SUMMARY OF PROVISIONS: Section 1: Subdivision 4 of section 205.00 of the penal law is amended to expand the definition of "dangerous contraband" to include firearms and other dangerous weapons, destructive devices and controlled substances.