Bill S2560 version original - Brad Hoylman-Sigal
BILL NUMBER: S2560 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the labor law, in relation to employee work schedules PURPOSE: This bill amends the Labor Law by adding a new Title 2 to Article 5 to create a responsibility upon certain employers to provide employees with consistent and predictable work schedules. SUMMARY OF SPECIFIC PROVISIONS: Section 1 creates a new Title 2 to Article 5 of the Labor Law and includes six subsections pursuant to the new Title.