Bill S2606 version original - Joseph A. Griffo
BILL NUMBER: S2606 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the state finance law, in relation to investment of moneys in the state fish and game trust account PURPOSE: To create a special account within the conservation fund to be known as the state fish and game trust account. SUMMARY OF PROVISIONS: Section 1.