Bill S2596 version original - Leroy Comrie
BILL NUMBER: S2596 SPONSOR: COMRIE TITLE OF BILL: An act to amend the insurance law, in relation to requirements for making of policy loans PURPOSE OR GENERAL IDEA OF BILL: The practice of churning has become increasingly prevalent in the life insurance market. In this practice, policyholders are encouraged to purchase additional life insurance policies by borrowing from the cash value of, or turning in, existing policies.