Bill S2525 version original - Pamela Helming
BILL NUMBER: S2525 SPONSOR: HELMING TITLE OF BILL: An act to amend the penal law, in relation to establishing a mandatory senior anti-violence services fee imposed upon conviction of a criminal offense and offenses against the elderly and disabled PURPOSE: To establish specific criminal offenses committed against the elderly or disabled and the creation of a senior anti-violence services fee. SUMMARY OF PROVISIONS: Sections 1: The section heading and subdivision 1 of section 60.