Bill S2584 version original - Kevin S. Parker
BILL NUMBER: S2584 SPONSOR: PARKER TITLE OF BILL: An act to amend the tax law, in relation to biofuel production credit for production of biomethane PURPOSE: Relates to production credit for production of biomethane. SUMMARY OF PROVISIONS: Section 1 renumbers section 38-a to Section 28 of the tax law, as added by section 1 of part X of chapter 62 of the laws of 2006.