Bill S2620 version original - Kevin S. Parker
BILL NUMBER: S2620 SPONSOR: PARKER TITLE OF BILL: An act to amend the public service law, in relation to prohibiting util- ities from imposing a moratorium on services in cities with a population of two hundred thousand or more PURPOSE: To halt unnecessary suspensions in service to both business and residen- tial customers SUMMARY OF PROVISIONS: Section 1. Subdivision 1 of section 65 of the public service law, is amended by chapter 789 of the law of 1930, is amended. Section 2.