Bill S2611 version original - Joseph A. Griffo
BILL NUMBER: S2611 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the environmental conservation law, in relation to lowering the age for universal hunting licenses from fourteen years old to twelve years old PURPOSE: The purpose of this bill is to lower the minimum age at which an indi- vidual may obtain universal hunting licenses. SUMMARY OF PROVISIONS: Section 1 amends subparagraph of paragraph b of subdivision 3 of section 11-0713 of the environmental conservation law.