Bill S2511 version original - Jessica Ramos
BILL NUMBER: S2511 SPONSOR: RAMOS TITLE OF BILL: An act to amend the executive law, in relation to defining employees of the state PURPOSE: This bill revises that law to define all workers of an elected state official as state employees, and, therefore, holds the state liable for ignoring complaints of workplace abuse. SUMMARY OF PROVISIONS: Section 1.