Bill S2590 version original - Leroy Comrie
BILL NUMBER: S2590 SPONSOR: COMRIE TITLE OF BILL: An act to amend the abandoned property law, in relation to extending the term for claiming unclaimed property from three to five years PURPOSE: To extend the time periods in the Abandoned Property Law for dormancy of account from three to five years. SUMMARY OF PROVISIONS: Section 1 amends the times periods of three years to five years for section 300 of the Abandoned Property Law.