Bill S2529 version original - Pamela Helming
BILL NUMBER: S2529 SPONSOR: HELMING TITLE OF BILL: An act to repeal certain provisions of chapter 1 of the laws of 2013, amending the criminal procedure law and other laws relating to suspen- sion and revocation of firearms licenses PURPOSE: This legislation would repeal chapter 1 of the laws of 2013 with the exception of sections 33 through 36. SUMMARY OF PROVISIONS: Section 1. Sections 1 through 32, 37 through 43, 45 through 56 of chap- ter 1 of the laws of 2013 are repealed. Section 2.