Bill S399B version A - Gustavo Rivera
BILL NUMBER: S399A SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law, in relation to enacting the safer consumption services act PURPOSE: This bill relates to the enactment of the Safer Consumption Services Act. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends Section 3399 subdivisions a-f of the Public Health law. Subdivision a is amended by defining "program" as an overdose prevention center instead of safer consumption services.