Bill S2522 version original - Pamela Helming
BILL NUMBER: S2522 SPONSOR: HELMING TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to creating the crime of stalking a police officer or peace officer PURPOSE: To provide additional protections to police officers and peace officers by creating the crime of stalking. SUMMARY OF PROVISIONS: Section 1: Amends Section 120.40 of the Penal Law to provide various definitions. Section 2: Adds a new Section 120.