Bill S2595 version original - Leroy Comrie
BILL NUMBER: S2595 SPONSOR: COMRIE TITLE OF BILL: An act to amend the education law, in relation to providing for the enhancement of the quality of teaching in the public schools in New York city PURPOSE OR GENERAL IDEA OF BILL: To address the looming shortage of qualified teachers willing to work in urban schools. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends the education law by adding a new article 125-C to establish the City University of New York Center for Excellence in Urban Education.