Bill S2598 version original - Leroy Comrie
BILL NUMBER: S2598 SPONSOR: COMRIE TITLE OF BILL: An act to amend the public authorities law, in relation to capital program plans PURPOSE OR GENERAL IDEA OF BILL: Relates to changing the date of submission of the capital program. plan to the capital program review board. SUMMARY OF PROVISIONS: Amends section 1269-b of public authorities law to add a new paragraph (d) and amend paragraphs (b) and (c).