Bill S2628 version original - Rachel May
BILL NUMBER: S2628 SPONSOR: MAY TITLE OF BILL: An act to amend the penal law, in relation to establishing incapacity to consent when a person is a witness to or subject of an investigation under certain circumstances PURPOSE: To establish incapacity to consent when a person is a witness or subject of an investigation under certain circumstances. SUMMARY OF PROVISIONS: Amends section 130.