Bill S2548 version original - Pamela Helming
BILL NUMBER: S2548 SPONSOR: HELMING TITLE OF BILL: An act to amend the penal law, in relation to sexual assault against a child by a person in a position of trust PURPOSE: To increase criminal penalties for sexual contact between a minor and a person in a position of trust. SUMMARY OF PROVISIONS: Article 130 of the NYS Penal Law is amended to add three new crimes in relation to sexual assault of a child by a person in a position of trust.