Bill S2587 version original - Kevin S. Parker
BILL NUMBER: S2587 SPONSOR: PARKER TITLE OF BILL: An act directing utilities to study the potential health risks and effects of smart meters; and providing for the repeal of such provisions upon expiration thereof PURPOSE: This bill will direct utilities and the public service commission to study the effects, and potential health risks associated with smart meters. SUMMARY OF PROVISIONS: 1.