Bill S2528 version original - Pamela Helming
BILL NUMBER: S2528 SPONSOR: HELMING TITLE OF BILL: An act to amend the criminal procedure law, in relation to the removal of a criminal action to a veterans treatment court PURPOSE: The purpose of the bill is to allow for the removal of a criminal action against a veteran to a veterans' court in the same or adjacent county. The measure only applies to counties outside New York City and removal may only be done with the consent of the court, the defendant and prose- cutor.