Bill S2640 version original - Leroy Comrie
BILL NUMBER: S2640 SPONSOR: COMRIE TITLE OF BILL: An act to amend the election law, in relation to additional days of voting as a result of emergencies PURPOSE: This bill would provide additional days of voting as a result of emer- gencies. SUMMARY OF PROVISIONS: Sections 1 and 2 of the bill set forth that this act shall be known as the "emergency response restoration of the electoral process act" and the legislative declaration.