Bill S2549 version original - Pamela Helming
BILL NUMBER: S2549 SPONSOR: HELMING TITLE OF BILL: An act to amend the criminal procedure law, in relation to providing that persons arrested for domestic violence shall forfeit bail when violating an order of protection PURPOSE: This bill would mandate an immediate forfeiture of bail when a full order of protection is violated. SUMMARY OF PROVISIONS: Section 1: amends the criminal procedure law to add a new section, 510.25.