Bill S2546 version original - Jessica Ramos
BILL NUMBER: S2546 SPONSOR: RAMOS TITLE OF BILL: An act to amend the workers' compensation law, in relation to notifica- tion in the case of a change to a workers' compensation claims represen- tative PURPOSE: To require insurance carriers to provide notice when a claims represen- tative changes or when it is determined that the assignment of a claims representative is no longer necessary. SUMMARY OF PROVISIONS: Section 1.