Bill S2646 version original - Mario Mattera
BILL NUMBER: S2646 SPONSOR: MATTERA TITLE OF BILL: An act to amend the tax law and the state finance law, in relation to gifts for the support of New York state small businesses PURPOSE: To allow taxpayers to contribute to the New York State Small Business Relief Fund SUMMARY OF PROVISIONS: Section 1 adds new section 630-1 to the tax law to read as followed. Gift for the New York state small business relief fund.