Bill S2538 version original - Leroy Comrie
BILL NUMBER: S2538 SPONSOR: COMRIE TITLE OF BILL: An act to direct the New York state department of health to conduct a study on the high incidence of asthma in the fifth congressional district in the city of New York and to prepare a remedial plan PURPOSE: To require the New York State Department of Health to conduct a year study on Asthma in Southeast Queens with a report and solutions to be given to the Governor and Legislature. SUMMARY OF PROVISIONS: § 1.