Bill S2520 version original - Anthony H. Palumbo
BILL NUMBER: S2520 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the education law, in relation to school district reor- ganizations and real property tax rates PURPOSE: To increase from 10 to 20 years the phase-in period in which the boards of education or trustees of school districts participating in a proposed reorganization may opt to have the tax impact of such reorganization calculated. SUMMARY OF PROVISIONS: Section 1.