Bill S2513 version original - Jessica Ramos
BILL NUMBER: S2513 SPONSOR: RAMOS PURPOSE OR GENERAL IDEA OF BILL: This bill repeals clause (D) of subparagraph (ii) of paragraph (a) of subdivision 1 of section 390 of the social services law relating to the definition of a child day care. SUMMARY OF SPECIFIC PROVISIONS: Section 1 This act shall be known and may be cited as "Isabella's Law". Section 2. Clause (D) of paragraph (ii) of paragraph (a) of subdivision 1 of section 3090 of the social services law is REPEATED. Section 2.