Bill S2667 version original - Leroy Comrie
BILL NUMBER: S2667 SPONSOR: COMRIE PURPOSE OR GENERAL IDEA OF BILL: Establishes the reasonable charge for electronic copies of medical records and patient information. SUMMARY OF SPECIFIC PROVISIONS: Both section 1 and 2, as amended by chapter 576 of the laws of 1998, is amended to add that the reasonable charge for copies provided in an electronic format shall be the lesser of either seventy-five cents per page or a total of one hundred dollars.