Bill S2760 version original - James Skoufis
BILL NUMBER: S2760 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the county law, in relation to abolishing the office of coroner PURPOSE: To require counties to end the archaic practice of electing an individ- ual to serve as county coroner. SUMMARY OF PROVISIONS: Section 1: Amends section 400 of the County Law to mandate that each county, other than Lewis and Madison County, shall abolish the office of coroner. Section 2: Sets effective date.