Bill S2804 version original - Andrew J. Lanza
BILL NUMBER: S2804 SPONSOR: LANZA TITLE OF BILL: An act to amend the public authorities law, in relation to the city of New York water board PURPOSE OF BILL: Relates to appointments to the water board of the city of New York. SUMMARY OF SPECIFIC PROVISIONS: Section one amends subdivision two of section 1045-f of the public authorities law, as added by chapter 513 of the laws of 1954. Section two is the effective date.