Bill S2829 version original - Andrew J. Lanza
BILL NUMBER: S2829 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to the crime of endangering the welfare of a child and to create a new crime of endangering the welfare of a child in the first degree PURPOSE OR GENERAL IDEA OF BILL: This bill creates the new crimes of endangering the welfare of a child. SUMMARY OF SPECIFIC PROVISIONS: Sections 1 through 6 amend Penal Law § 70.02, 260.10, 260.12, and 260.15, renumber current Penal Law § 260.11 as § 260.