Bill S2809 version original - Andrew J. Lanza
BILL NUMBER: S2809 SPONSOR: LANZA TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to the declaration of a sign as a public nuisance PURPOSE OR GENERAL IDEA OF BILL: To remove billboards that have been found to violate the New York zoning resolution, as well as the structures attached to them. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Amends Section 28-503.