Bill S2828 version original - Andrew J. Lanza
BILL NUMBER: S2828 SPONSOR: LANZA TITLE OF BILL: An act to amend the criminal procedure law and the family court act, in relation to notification of service of ex parte orders of protection PURPOSE: Provides for the immediate notification to a complainant, in the manner specified by the complainant, of the service of an ex parte order of protection. SUMMARY OF PROVISIONS: Section 1 The criminal procedure law is amended by adding a new section 530.