Bill S2823 version original - Andrew J. Lanza
BILL NUMBER: S2823 SPONSOR: LANZA TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the regis- tered weight of tow trucks PURPOSE OR GENERAL IDEA OF BILL: To increase the minimum weight of a tow truck from 8,600 pounds GVWR (gross vehicle weight rating) to 10,000 pounds GVWR SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends section 401-b of the vehicle and traffic law to increase the minimum GVWR of a tow truck from 8,000 pounds to 10,000 pounds. Section 2.