Bill S2676A version original - Nathalia Fernandez
BILL NUMBER: S2676 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the public health law and the environmental conservation law, in relation to indoor air quality in schools PURPOSE OR GENERAL IDEA OF BILL: To establish an indoor air quality program in schools SUMMARY OF PROVISIONS: Section 1 adds a new title (4-a) to public health law regarding indoor air quality in schools. Section 2 requires the DEC to cooperate with the public health depart- ment. Section 3 states the effective date.