Bill S2789 version original - James Skoufis
BILL NUMBER: S2789 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the penal law, in relation to creating the crime of aggravated harassment of a survivor of domestic violence or sexual assault PURPOSE: To better protect survivors of domestic violence and sexual abuse from cyberbullying. SUMMARY OF PROVISIONS: Section 1: Amends the Penal Law to add a new section 240.33 Aggravated harassment of a survivor of domestic violence or sexual assault.