Bill S2692 version original - Neil D. Breslin
BILL NUMBER: S2692 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the county law, in relation to reimbursement for coroner and medical examiner fees from the county where the decedent resided PURPOSE OR GENERAL IDEA OF BILL: To amend the county law, in relation to reimbursement for coroner and medical examiner fees from the county where the decedent resided. SUMMARY OF PROVISIONS: Section 1. The county law is amended by adding a new section 679. Section 2.