Bill S2717 version original - Rachel May
BILL NUMBER: S2717 SPONSOR: MAY TITLE OF BILL: An act to amend the civil practice law and rules, in relation to author- izing a program for electronic filing for commencement of civil actions and proceedings and service of papers in town and village courts PURPOSE: To permit electronic filing of civil actions and proceedings and service of papers in town and village courts. SUMMARY OF PROVISIONS: Section 1 adds town and village courts to the list of courts that may accept electronic filing.