Bill S2813 version original - Neil D. Breslin
BILL NUMBER: S2813 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to coverage for certain benefits under comprehensive motor vehicle insurance PURPOSE: The purpose of this legislation is to eliminate the assignment of bene- fits for durable medical equipment in the no-fault system. SUMMARY OF PROVISIONS: Section 1 adds a new sub Section (n) to section 5102 of the insurance law to define "health service provider.