Bill S2794 version original - Neil D. Breslin
BILL NUMBER: S2794 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to establishing the interstate insurance product regulation compact to regulate certain insurance products PURPOSE: Establishes an interstate insurance product regulation compact. The purposes of this compact are, through means of joint and cooperative action among the compacting states: a.