Bill S2807 version original - Andrew J. Lanza
BILL NUMBER: S2807 SPONSOR: LANZA TITLE OF BILL: An act to amend the correction law, in relation to in-person appearance requirements for level two sex offenders PURPOSE: Increases the in-person appearance requirements for level two sex offen- ders. SUMMARY OF PROVISIONS: The correction law is amended by adding section 2 of part 0 of chapter 55.