Bill S2808 version original - Andrew J. Lanza
BILL NUMBER: S2808 SPONSOR: LANZA TITLE OF BILL: An act to amend the environmental conservation law, in relation to prohibiting the disposal of dredged spoils containing toxic pollutants into the waters of the marine district PURPOSE: Prohibits disposal of spoils containing toxic pollutants into the waters of the marine district. SUMMARY OF PROVISIONS: This legislation amends .5 17-0503 of the Environmental Conservation Law by adding a new subdivision 3.