Bill S2787 version original - James Skoufis
BILL NUMBER: S2787 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the general construction law and the administrative code of the city of New York, in relation to observing Atlantic time PURPOSE: To move New York state to Atlantic Standard Time, which would mean later sunsets in the winter. SUMMARY OF PROVISIONS: Section 1: Directs the secretary of state to develop and submit an application to the federal government requesting that New York observe Atlantic Standard Time.