Bill S2759 version original - James Skoufis
BILL NUMBER: S2759 SPONSOR: SKOUFIS TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing an amendment to section 11 of article 1 of the constitution, in relation to adding disability to the equal protection requirement PURPOSE: Adds the term "disability" to the New York State Constitution's protections against discrimination. SUMMARY OF PROVISIONS: Amends section 11 of article 1 of the Constitution to add the term "disability.