Bill S2686 version original - Neil D. Breslin
BILL NUMBER: S2686 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the environmental conservation law, in relation to prohibiting the use of ash from burning hazardous waste in products SUMMARY OF SPECIFIC PROVISIONS: Section 1 describes the legislative intent. Section 2 adds a new section to the environmental conservation law, section 19-0304-a. Section 3 is a severability clause. Section 4 is the effective date.