Bill S2749A version original - Jamaal Bailey
BILL NUMBER: S2749 SPONSOR: BAILEY TITLE OF BILL: An act to amend the state finance law, in relation to establishing the New York state youth sports initiative grants fund PURPOSE: The purpose of this legislation is to establish a fund to promote youth sports across our state. SUMMARY OF SPECIFIC PROVISIONS: The first section of the bill establishes the "youth sports initiative grants fund" under the direction of the state comptroller and the commissioner of education.