Bill S2740 version original - Leroy Comrie
BILL NUMBER: S2740 SPONSOR: COMRIE TITLE OF BILL: An act to amend the state finance law, in relation to the purchase of green cleaning and other custodial products PURPOSE: To improve the indoor air quality of state owned and occupied buildings to protect the individuals who work and visit from inhaling toxic and harmful chemicals. SUMMARY OF PROVISIONS: Section 1 titles the bill the "Greening Our Cleaning Act." Section 2 sets forth a statement of findings and purpose.