Bill S2778 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S2778 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the criminal procedure law, in relation to designating sheriffs, under-sheriffs and deputy sheriffs of the city of New York as police officers PURPOSE: Relates to the definition of a police officer. SUMMARY OF PROVISIONS: Section 1 amend- paragraph (o) of subdivision 34 of section 1.20 of the Criminal Procedure Law.