Bill S2721 version original - Rachel May
BILL NUMBER: S2721 SPONSOR: MAY TITLE OF BILL: An act to amend the executive law, the real property actions and proceedings law and the real property law, in relation to establishing the New York state office of civil representation to provide a right to counsel in eviction proceedings PURPOSE: Establishes the civil right to counsel in eviction proceedings in New York State. SUMMARY OF PROVISIONS: Section 1: Sets forth legislative findings.