Bill S2822 version original - Anthony H. Palumbo
BILL NUMBER: S2822 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the mental hygiene law, in relation to proceedings for appointment of a guardian or other caretaker for personal needs or prop- erty management PURPOSE: To expedite hearings on the guardianship/appeals process for family visitation. SUMMARY OF PROVISIONS: Section I. Cites this act as "Karilyn's Law. Section 2. Amends section 81.