Bill S2719 version original - Rachel May
BILL NUMBER: S2719 SPONSOR: MAY TITLE OF BILL: An act to amend the uniform justice court act, in relation to raising the jurisdictional amount for small claims cases PURPOSE: To raise the jurisdictional amount that may be brought before a small claims court. SUMMARY OF PROVISIONS: Section 1 amends section 201 of the uniform court justice act to raise the cap for claims that may be prosecuted in small claims court to $10,000.