Bill S2766 version original - Leroy Comrie
BILL NUMBER: S2766 SPONSOR: COMRIE TITLE OF BILL: An act to amend the general obligations law, in relation to the provisions of consumer and employment contracts PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to clarify basic precepts in law concerning contract formation. SUMMARY OF SPECIFIC PROVISIONS: This bill would give consumers and employees the right to reform contracts which are not written in plain language or which are not whol- ly encompassed within a single document.