Bill S2675 version original - Nathalia Fernandez
BILL NUMBER: S2675 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the executive law, in relation to establishing August thirty-first of each year as a day of commemoration known as "Overdose Awareness Day" PURPOSE OR GENERAL IDEA OF BILL: To designate August 31st as Overdose Awareness Day SUMMARY OF PROVISIONS: Section 1 adds Overdose Awareness Day as a day of commemoration on August 31. Section 2 establishes the effective date. JUSTIFICATION: Overdose death is preventable.