Bill S2715 version original - Patrick M. Gallivan
BILL NUMBER: S2715 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the civil practice law and rules, in relation to enact- ing the court order protection act PURPOSE OF BILL: To clarify a court's responsibility in the case of judgment entry and a debtor's responsibility in the case of a court ordered payment. SUMMARY OF SPECIFIC PROVISIONS: Section 1. establishes the "court order protection act.