Bill S2925 version original - Cordell Cleare
BILL NUMBER: S2925 SPONSOR: CLEARE TITLE OF BILL: An act to amend the mental hygiene law, in relation to ensuring that the office of mental health and the office for people with developmental disabilities provide materials to individuals in different languages PURPOSE: To require the Office for Mental Health and the Office for People With Developmental Disabilities to have materials printed in English as well as other languages. SUMMARY OF PROVISIONS: Section 1 amends section 7.