Bill S2905 version original - Dan Stec
BILL NUMBER: S2905 SPONSOR: STEC TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as the "Lance Cpl. Scott Lee Schultz Memorial Bridge" PURPOSE: To dedicate a bridge on a state highway to Lance Cpl. Scott Lee Schultz SUMMARY OF PROVISIONS: Section 1 - Designates an overpass on state Route 22, in the town of Chesterfield, the "Lance Cpl.