Bill S2884 version original - Leroy Comrie
BILL NUMBER: S2884 SPONSOR: COMRIE TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring certain motor vehicle forms and applications to be available in Punjabi, Bengali and Hindi PURPOSE: The purpose of this bill is to require the Department of Motor Vehicles to provide forms, tests and applications for permits and driver's licenses in Punjabi, Bengali, & Hindi. Translation and interpretation services will also be made available upon request.