Bill S2941 version original - Pamela Helming
BILL NUMBER: S2941 SPONSOR: HELMING TITLE OF BILL: An act to amend the correction law, in relation to creating a contraband and drug information hotline PURPOSE: To provide a resource for people to report any knowledge about drugs or weapons in our prison systems that may be of harm to correction offi- cers, inmates, and visitors. SUMMARY OF PROVISIONS: Sections 1: The corrections law is amended to provide a definition for contraband.