Bill S568B version A - Brian Kavanagh
BILL NUMBER: S568A SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the public housing law, in relation to establishing the housing access voucher program PURPOSE: To provide housing assistance to the homeless and those facing an immi- nent risk of homelessness. SUMMARY OF SPECIFIC PROVISIONS: This bill creates the Housing Access Voucher Program, which provides rental assistance in the form of vouchers to individuals who are home- less or who face an imminent risk of losing their homes.