Bill S2838 version original - Neil D. Breslin
BILL NUMBER: S2838 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the education law, in relation to the New York state masters-in-education teacher incentive scholarship program PURPOSE: This bill will align the Masters-in-Education Teacher Incentive Scholar- ship Program with other state scholarship programs allowing students to choose the college or university with the academic program and approach that benefits them most. SUMMARY OF PROVISIONS: Section 1.