Bill S2935D version original - Kevin S. Parker
BILL NUMBER: S2935 SPONSOR: PARKER TITLE OF BILL: An act relating to the development of a blueprint to guide the replace- ment and redevelopment of New York's oldest and most-polluting fossil fuel facilities and their sites by 2030 PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to provide a blueprint to guide the replace- ment and redevelopment of New York State's fossil fuel facilities and their sites by 2030. SUMMARY OF PROVISIONS: Section 1 identifies the title of the bill.